move `Support` under the `Desktop` namespace
