+++ src/Desktop/Drivers/Electron/Commands/BuildCommand.php
@@ -5,12 +5,12 @@ namespace Native\Desktop\Drivers\Electron\Commands;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Process;
 use Illuminate\Support\Str;
+use Native\Desktop\Builder\Builder;
 use Native\Desktop\Drivers\Electron\ElectronServiceProvider;
 use Native\Desktop\Drivers\Electron\Facades\Updater;
 use Native\Desktop\Drivers\Electron\Traits\InstallsAppIcon;
 use Native\Desktop\Drivers\Electron\Traits\OsAndArch;
 use Native\Desktop\Drivers\Electron\Traits\PatchesPackagesJson;
-use Native\Support\Builder;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Process\Process as SymfonyProcess;
