+++ src/Desktop/Drivers/Electron/Commands/BundleCommand.php
@@ -9,10 +9,10 @@ use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Process;
 use Illuminate\Support\Number;
 use Illuminate\Support\Str;
+use Native\Desktop\Builder\Builder;
 use Native\Desktop\Drivers\Electron\Traits\HandlesZephpyr;
 use Native\Desktop\Drivers\Electron\Traits\InstallsAppIcon;
 use Native\Desktop\Drivers\Electron\Traits\PatchesPackagesJson;
-use Native\Support\Builder;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Finder\Finder;
 use ZipArchive;
