+++ src/Desktop/Drivers/Electron/Commands/DevelopCommand.php
@@ -3,12 +3,12 @@
 namespace Native\Desktop\Drivers\Electron\Commands;
 
 use Illuminate\Console\Command;
+use Native\Desktop\Builder\Builder;
 use Native\Desktop\Drivers\Electron\ElectronServiceProvider;
 use Native\Desktop\Drivers\Electron\Traits\Developer;
 use Native\Desktop\Drivers\Electron\Traits\Installer;
 use Native\Desktop\Drivers\Electron\Traits\InstallsAppIcon;
 use Native\Desktop\Drivers\Electron\Traits\PatchesPackagesJson;
-use Native\Support\Builder;
 use Symfony\Component\Console\Attribute\AsCommand;
 
 use function Laravel\Prompts\intro;
